+++ docs/docs/changelog.md
@@ -13,8 +13,9 @@ title: Changelog
 * [Refactoring] Refactored `runner` package, implemented `Runner` struct.
 * [Added] Support `NO_COLOR` env variable to disable colored output. See https://no-color.org/
 * [Added] `LETS_COMMAND_ARGS` - will contain command's positional args. [See config](/docs/env#default-environment-variabless).
+  
   Also, special bash env variables such as `"$@"` and `"$1"` etc. now available inside `cmd` script and work as expected. 
-
+* [Added] `work_dir` directive. See [config](/docs/config#work_dir)
   
 ## [0.0.33](https://github.com/lets-cli/lets/releases/tag/v0.0.33)